requirements.txt
pyQt5
Pillow
twisted